refactor(helpers): replaced folder name and separate files
