+++ src/pages/ui-components/Typography/index.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import styled, { css, IStyledProp } from '@eduzz/houston-styles/styled';
 import type { HoustonTokens } from '@eduzz/houston-tokens';
 
-import { getColorFallback } from '../Helpers/functions';
+import getColorFallback from '../utils/getColorFallback';
 
 export type TypographyColors = 'low' | 'high' | 'primary';
